Add files via upload
We are examining the data and figuring out the level of the variable whether it is categorical or numerical.
We are putting each variable in a dataframe so then we can utilize it to show visual and statistical representation
of the data. We show this via different types of graphs and statisitical components such as the mean and correlation.
Our graphs compare our different target variables like industry, revenue, number of employees... to eachother to better
get a sense of how one variable relates to another. Ultimately, we are taking the data of the top 100 companies in the
United States to compare them and give some analytical context of the data.